Migrate from coq-prelude to coq-ext-lib
Since its initial release, coq-comparse was using coq-prelude, a
home-grown, “opinionated” prelude library for Coq. This library was
published as an open source project, but never made available as an
OPAM package. In the meantime, the ExtLib coq project became more and
more popular, and often exceeds the scope of coq-prelude.

As a consequence, we take the decision to move away from coq-prelude,
in order to adopt coq-ext-lib instead.